refactor: controller 구조 변경 및 app/ 경로 처리
