+++ src/main/java/io/github/daengdaenglee/mangurl/inboundadapter/shorturl/ShortUrlController.java
@@ -1,4 +1,4 @@
-package io.github.daengdaenglee.mangurl.inboundadapter.root;
+package io.github.daengdaenglee.mangurl.inboundadapter.shorturl;
 
 import io.github.daengdaenglee.mangurl.application.url.inboundport.EncodeUrlService;
 import io.github.daengdaenglee.mangurl.application.url.inboundport.RestoreUrlService;
@@ -13,7 +13,7 @@ import org.springframework.web.server.ResponseStatusException;
 @Slf4j
 @RequiredArgsConstructor
 @Controller
-class RootController {
+class ShortUrlController {
     private final RestoreUrlService restoreUrlService;
     private final EncodeUrlService encodeUrlService;
 
@@ -23,10 +23,5 @@ class RootController {
                 .map(this.encodeUrlService::encode)
                 .map(url -> "redirect:" + url)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
-    }
-
-    @RequestMapping({"", "/"})
-    String home() {
-        return "redirect:/app";
     }
 }
